+++ modules/kubernetes/grafana-agent/variables.tf
@@ -47,11 +47,17 @@ variable "vpa_enabled" {
 }
 
 variable "namespace_include" {
-  description = "A list of the namespaces that kube-state-metrics should create metrics for"
+  description = "A list of the namespaces that kube-state-metrics and kubelet metrics"
   type        = list(string)
 
   validation {
     condition     = length(var.namespace_include) > 0
     error_message = "The namespace_include needs to at least contain one namespace in the list."
   }
+}
+
+variable "include_kubelet_metrics" {
+  description = "If kubelet metrics shall be included for the namespaces in 'namespace_include'"
+  type        = bool
+  default     = false
 }