+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - [#871](https://github.com/XenitAB/terraform-modules/pull/871) Change ingress-nginx affinity match labels.
 
+### Added
+
+- [#857](https://github.com/XenitAB/terraform-modules/pull/857) Grafana agent kubelet metrics.
+
 ## 2022.11.2
 
 ### Added
@@ -20,6 +24,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Fixed
 
 - [#865](https://github.com/XenitAB/terraform-modules/pull/865) Make azad-kube-proxy AD group filter more specific.
+- [#873](https://github.com/XenitAB/terraform-modules/pull/873) Add linkerd exception to default-deny networkpolicy.
+- [#874](https://github.com/XenitAB/terraform-modules/pull/874) Update linkerd to 2.12.2.
 
 ## 2022.11.1